cuda image: int->byte

+++ c-fractals/src/cmatrix.c
@@ -147,7 +147,7 @@ fractal_image_save(uint8_t * image, int width, int height, const char * filename
 uint8_t *
 fractal_image_create(int height, int width)
 {
-    return (uint8_t *) malloc(sizeof(int)*height*width*3);
+    return (uint8_t *) malloc(sizeof(uint8_t)*height*width*3);
 }
 
 void
